+++ front/src/layouts/ExperimentConfiguration.jsx
@@ -1,15 +1,16 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import {
   Container,
   Row,
   Col,
+  // Dropdown,
+  // DropdownButton,
   Form,
-  Dropdown,
-  DropdownButton,
+  Button,
 } from 'react-bootstrap';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-// import ModelsTable from '../components/ModelsTable';
+import ModelsTable from '../components/ModelsTable';
 import Upload from '../components/Upload';
 import ParameterForm from '../components/ParameterForm';
 
@@ -17,27 +18,44 @@ const StyledContainer = styled(Container)`
   margin: 20px;
 `;
 
-function SelectModels({ availableModels, checkedOptions, setCheckedOptions }) {
-  SelectModels.propTypes = {
+function AddModels({
+  availableModels,
+  modelsInTable,
+  setModelsInTable,
+  setParameterSchema,
+}) {
+  AddModels.propTypes = {
     availableModels: PropTypes.arrayOf(PropTypes.string).isRequired,
-    checkedOptions: PropTypes.arrayOf(PropTypes.string).isRequired,
-    setCheckedOptions: PropTypes.func.isRequired,
+    modelsInTable: PropTypes.arrayOf(PropTypes.objectOf(PropTypes.string)).isRequired,
+    setModelsInTable: PropTypes.func.isRequired,
+    setParameterSchema: PropTypes.func.isRequired,
   };
-  const handleChange = ({ target }) => {
-    if (target.checked) {
-      setCheckedOptions([...checkedOptions, target.name]);
-    } else {
-      setCheckedOptions(checkedOptions.filter((x) => x !== target.name));
-    }
+  const [addModelValues, setAddModelValues] = useState({ name: '', type: '' });
+  const handleSubmit = (e) => {
+    e.preventDefault(e);
+    setModelsInTable([...modelsInTable, addModelValues]);
+  };
+  const handleChange = (e) => {
+    setAddModelValues((state) => ({
+      ...state,
+      [e.target.name]: e.target.value,
+    }));
   };
   if (availableModels.length !== 0) {
     return (
       <div>
         <h4>Task Type: Text Classification</h4>
-        <p>Select the models to train.</p>
-        <Form>
-          {availableModels.map((model) => <Form.Check inline onChange={handleChange} label={model} name={model} type="checkbox" key={`checkbox-${model}`} />)}
+        <p>Add models to train.</p>
+        <Form className="d-flex" style={{ display: 'grid', gridGap: '10px' }}>
+          <input type="text" name="name" value={addModelValues.name} onChange={handleChange} />
+          <select value={addModelValues.type} name="type" onChange={handleChange}>
+            <option>Select model</option>
+            { availableModels.map((model) => <option value={model} key={model}>{model}</option>) }
+          </select>
+          <Button onClick={handleSubmit}>Add</Button>
         </Form>
+        <br />
+        <ModelsTable rows={modelsInTable} setParameterSchema={setParameterSchema} />
       </div>
     );
   }
@@ -45,62 +63,26 @@ function SelectModels({ availableModels, checkedOptions, setCheckedOptions }) {
 }
 
 function ExperimentConfiguration() {
-  const [response, setResponse] = useState({ knn: { } });
   const [availableModels, setAvailableModels] = useState([]);
-  const [selectedModels, setSelectedModels] = useState([]);
-  const [configOption, setConfigOption] = useState('Select model');
-  const [parameterSchema, setParameterSchema] = useState({});
-  const handleSelect = async (eventkey, event) => {
-    const selectedOption = event.target.firstChild.data;
-    setConfigOption(selectedOption);
-    const fetchedOption = await fetch(`http://localhost:8000/selectModel/${selectedOption}`);
-    const formJson = await fetchedOption.json();
-    setParameterSchema(formJson);
-  };
-  useEffect(() => {
-    async function fetchData() {
-      const fetched = await fetch('http://localhost:8000/experiment/results/0');
-      const res = await fetched.json();
-      setResponse(res);
-    }
-    fetchData();
-  }, []);
+  const [modelsInTable, setModelsInTable] = useState([]);
+  const [parameterSchema, setParameterSchema] = useState(null);
   return (
     <StyledContainer>
       <Row>
         <Col md="6">
           <h2>Load Dataset</h2>
           <Upload setModels={setAvailableModels} />
-          <SelectModels
+          <AddModels
             availableModels={availableModels}
-            checkedOptions={selectedModels}
-            setCheckedOptions={setSelectedModels}
+            modelsInTable={modelsInTable}
+            setModelsInTable={setModelsInTable}
+            setParameterSchema={setParameterSchema}
           />
-          {/* <ModelsTable /> */}
         </Col>
 
         <Col md="6">
-          {
-            selectedModels.length > 0
-            && (
-            <div>
-              <DropdownButton title={configOption} variant="secondary" onSelect={handleSelect}>
-                {
-                  selectedModels.map(
-                    (model) => <Dropdown.Item key={model}>{model}</Dropdown.Item>,
-                  )
-                }
-              </DropdownButton>
-              <br />
-              {configOption !== 'Select model'
-                  && <ParameterForm model={configOption} parameterSchema={parameterSchema} />}
-            </div>
-            )
-          }
-          <p>
-            Model Accuracy:
-            {response.knn.accuracy}
-          </p>
+          { parameterSchema !== null
+          && <ParameterForm model="" parameterSchema={parameterSchema} />}
         </Col>
       </Row>
     </StyledContainer>
